src/agent/db/connection.py
import psycopg2
from contextlib import contextmanager
from agent.config import settings

@contextmanager
def get_conn():
    conn = psycopg2.connect(
        host=settings.db_host,
        port=settings.db_port,
        dbname=settings.db_name,
        user=settings.db_user,
        password=settings.db_password,
    )
    try:
        yield conn
    finally:
        conn.close()
